Began implementation of mobilenetv3 and vgg, as well as fixing some layer freezing and thawing (also broke the functionality).
