+++ src/custom_image_dataset.py
@@ -23,6 +23,7 @@ class CustomImageDataset(Dataset):
 
     def __getitem__(self, idx):
         img_path = os.path.join(self.img_dir, self.img_labels.iloc[idx, 0])
+        print(img_path)
         image = read_image(img_path)[:3, :, :]
         label = self.img_labels.iloc[idx, 1]
         if self.transform:
