+++ src/helper_functions.py
@@ -1,42 +1,42 @@
 import os
 from datetime import datetime
 import csv
+from typing import Optional
 
-from matplotlib import pyplot as plt
 from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
 import matplotlib
 
+from logging_level import LoggingLevel
+
 matplotlib.use('TkAgg')
 
-color_codes = {
-    'DEBUG': 'lightsteelblue',
-    'INFO': 'white',
-    'WARNING': 'yellow',
-    'ERROR': 'red'
-}
 logger = None
 static_window = None
 static_canvas = None
 static_fig_agg = None
 
 
-def log_message(message: str, level: str = 'INFO', window=None):
+def log_message(message: str, level: Optional[LoggingLevel] = LoggingLevel.INFO, window=None):
     global logger
     global static_window
 
+    # Build the message
     ts = datetime.now()
     log_msg = f'{ts} | '
-    log_msg += ' %7s | ' % level
+    log_msg += ' %7s | ' % level.name
     log_msg += message + '\n'
 
+    # If a PySimpleGUI window is passed through, then set that window as the static_window to update
     if window is not None:
         logger = window['output_box'].Widget
         static_window = window
+    # If static_window doesn't exist by this point in the code, then it is being run without the GUI and we should
+    # simply print the logging message.
     if static_window is None:
         print(log_msg)
         return
 
-    logger.tag_config(level, background='black', foreground=color_codes[level])
+    logger.tag_config(level, background='black', foreground=level.value)
 
     logger.insert('end', log_msg, level)
     static_window.refresh()
