+++ src/main.py
@@ -1,0 +1,3 @@
+import user_interface
+
+user_interface.run_gui()
